+++ src/main/java/entity/Portfolio.java
@@ -12,26 +12,8 @@ public class Portfolio {
     }
 
     // Constructor that initializes Portfolio with a map of tickers and their respective quantities
-    public Portfolio(Map<String, Integer> stocks) {
-        this.stocks = new HashMap<>();
-        StockMarket stockMarket = StockMarket.Instance();
-
-        stocks.forEach((ticker, quantity) -> {
-            // decompose Optional<Stock>
-            stockMarket.getStock(ticker).ifPresentOrElse(
-                stock -> {
-                    double price = stock.getPrice();
-                    if (price > 0) {
-                        this.stocks.put(ticker, new UserStock(stock, price, quantity));
-                    } else {
-                        throw new IllegalArgumentException("Invalid price for stock: " + ticker);
-                    }
-                },
-                () -> {
-                    throw new IllegalArgumentException("No stock found for ticker: " + ticker);
-                }
-            );
-        });
+    public Portfolio(Map<String, UserStock> stocks) {
+        this.stocks = new HashMap<>(stocks);
     }
 
     public double getTotalValue() {