+++ docs/CHANGELOG.md
@@ -13,5 +13,6 @@
 - Added Use Cases 4 and 5.
 - Implemented entities for stock and stock market.
 - Implemented entities for user, portfolio, and transaction.
-- Add synchronization checks and null checks in StockMarket
+- Add synchronization checks and null checks in `StockMarket`
 - Modify `Portfolio` constructor to match the optional return value of `StockMarket.getStock`
+- Modify `Portfolio` constructor to take `Map<String, UserStock>` as parameter and remove synchronization with `StockMarket` data